feature/onboarding (#176)
* [Feat] Main -> Setting 커스텀 슬라이드 애니메이션 추가

* [Design] TextButton 구현, SignUpVIewController onboardingButton 구현

* [refactor] SignUpViewController TransitionStyle 추가

* [Refactor] UserDefaultManager isOnboarding 추가 및 UIViewController+Extension 수정

* [Feat] OnboardingAPI 및 MainViewController 최근 임장 불러오기 구현

* [feat] 회원가입유도 바텀시트 구현

* [refactor] 설정뷰 온보딩 분기 UI 및 기능 구현

* [feat] 새노트 생성 온보딩 분기처리 추가

* [Chore] 온보딩 관련 분기처리 주석 추가

* [refactor] mainVC getProfileInfo 온보딩 분기처리 추가

* [Fix] 로그인 시도후 실패 시 닉네임 변경 이슈 해결

* [feat] 나의 임장 노트 뷰 온보딩 분기처리 추가

* [feat] 나의 임장 노트 검색 온보딩 분기처리 추가

* [feat] 나의 임장노트 온보딩 GET 노트 구현

* [chore] 임장 리스트 뷰 dependency 추가

* [refactor] 임장 상세 pushViewController 시 온보딩 분기 추가

* [feat] SignUpToastView 구현 , 임장 상세 온보딩 분기처리 추가

* [Feat] 나의 임장노트 온보딩 체크리스트 API 연동

* [refactor] 임장이미지뷰컨 이미지 더하기, 삭제하기 온보딩 로직 추가

* [refactor] 나의 임장 상세 디테일, 공유조건 데이터, 메모장, 녹음 온보딩 분기처리

* [feat] 임장 상세 레코딩 룸 온보딩 API 연결

* [feat] 리포트 온보딩 API 연동

* [refactor] mynotes path 수정

* [refactor] onboardingAPI recentnotes 네이밍 변경

* [fix] 메인뷰 바텀 로딩뷰 무한 로딩 해결

* [fix] 마이노트 상세 공유 조건 뷰 안보이는 현상 수정

* [fix] tripleImageView 로직 수정

* [Fix] ImjangImage 이미지 안보이는 현상 수정

* [fix] 공유 하기 버튼 다중 클릭 수정

* [fix] scrapCell 이미지간 여백 수정

* [fix] 바텀 시트 여백 수정

* [refactor] 마이노트 검색 온보딩 마이노트 API로 구현

* [fix] pencil 이용약관 API 온보딩 시 요청 X

* [fix] settingViewController getProfileInfo 온보딩 로직 추가

* [fix] 전체보기 클릭시 바텀시트

* [fix] 공유하기 클릭시 API 호출 X

* [fix] 리포트 온보딩 모드 시 API 콜 X

* [fix] 체크리스트 text answer background 안보임 수정

* [chore] appVersion 수정

* [refactor] OnboardingAPI 코드 리팩

* [chore] delete Untitled 2.swift

* [Chore] delete Untitled.swift

* [fix] 임장 상세 도움말 팝업 안뜨는 현상 수정

* feat: 빌드넘버 업데이트

* [fix] 로그인이 되어 있는 경우 isOnboading false

* [Chore] dev 빌드 버전 변경

---------

+++ juinjang/Projects/App/Sources/UserInterface/ImjangList/ImjangListViewController.swift
@@ -48,6 +48,7 @@ final class ImjangListViewController: BaseViewController {
     
     struct Dependency {
         let noteRepository: NoteRepositoryProtocol
+        let onboardingRepository: OnboardingRepositoryProtocol
     }
     
     private let dependency: Dependency
@@ -103,8 +104,32 @@ final class ImjangListViewController: BaseViewController {
 // MARK: - request
 extension ImjangListViewController {
     private func fetchImjangList(sort: MyNoteFilter = .updated, setScrap: Bool = false) {
-        print(#function)
         setLoading(isShow: true)
+
+        if UserDefaultManager.shared.isOnboarding {
+            dependency
+                .onboardingRepository
+                .retrieveMyNotes()
+                .asObservable()
+                .catch { [weak self] error in
+                    self?.setLoading(isShow: false)
+                    self?.showAlert(title: "에러", message: error.localizedDescription, actionHandler: nil)
+                    return .empty()
+                }
+                .subscribe(with: self) { (self, response) in
+                    self.mainView.setupEmptyView(isEmpty: response.isEmpty)
+                    self.imjangList = response
+                    self.setData(scrapedList: response)   // 스크랩된것들 scrapList에 추가
+                    self.mainView.hasResults(!self.imjangList.isEmpty)
+                    self.mainView.collectionView.reloadData()
+                    self.setLoading(isShow: false)
+                }
+                .disposed(by: disposeBag)
+            
+            return
+        }
+        
+        print(#function)
         dependency
             .noteRepository
             .retrieveNoteList(sort: sort.parameterValue, keyword: "")
@@ -187,6 +212,12 @@ extension ImjangListViewController: DeleteImjangListDelegate {
     
     // 삭제 화면으로 이동
     @objc private func showDeleteImjangVC() {
+        // 온보딩 분기처리
+        if UserDefaultManager.shared.isOnboarding {
+            present(SignUpBottomSheetView(), animated: true)
+            return
+        }
+        
         let DeleteImjangVC = DeleteImjangViewController(
             dependency: DeleteImjangViewController.Dependency(
                 noteRepository: NoteRepository()
@@ -197,6 +228,12 @@ extension ImjangListViewController: DeleteImjangListDelegate {
     }
     
     @objc private func showShareSelectVC() {
+        // 온보딩 분기처리
+        if UserDefaultManager.shared.isOnboarding {
+            present(SignUpBottomSheetView(), animated: true)
+            return
+        }
+        
         let viewController = ShareSelectViewController(
             reactor: .init(
                 dependency: .init(
@@ -273,6 +310,9 @@ extension ImjangListViewController: DeleteImjangListDelegate {
     }
 
     private func scrapRequest(imjangId: Int) {
+        // 온보딩 분기처리
+        guard (!UserDefaultManager.shared.isOnboarding) else { return }
+
         JuinjangAPIManager.shared.fetchData(type: NoResultResponse.self,
                                             api: .scrap(imjangId: imjangId)) { response, error in
             if let error = error {
@@ -285,6 +325,9 @@ extension ImjangListViewController: DeleteImjangListDelegate {
     }
         
     private func cancelScrapRequest(noteId: Int) {
+        // 온보딩 분기처리
+        guard (!UserDefaultManager.shared.isOnboarding) else { return }
+        
         JuinjangAPIManager.shared.fetchData(type: NoResultResponse.self, api: .cancelScrap(imjangId: noteId)) { response, error in
             if let error = error {
                 print(error.localizedDescription)
@@ -354,7 +397,8 @@ extension ImjangListViewController: DeleteImjangListDelegate {
 
 // MARK: - CollectionView Delegate
 extension ImjangListViewController: UICollectionViewDataSource, UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+    func collectionView(_ collectionView: UICollectionView,
+                        numberOfItemsInSection section: Int) -> Int {
         if let imjangSection = Section(rawValue: section) {
             switch imjangSection {
             case .scrap:
@@ -400,11 +444,19 @@ extension ImjangListViewController: UICollectionViewDataSource, UICollectionView
         }
     }
     
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+    func collectionView(_ collectionView: UICollectionView,
+                        didSelectItemAt indexPath: IndexPath) {
         if let imjangSection = Section(rawValue: indexPath.section) {
             switch imjangSection {
             case .scrap:
                 let item = scrapImjangList[indexPath.row]
+                
+                // 온보딩 분기처리
+                if UserDefaultManager.shared.isOnboarding {
+                    self.showImjangNoteVC(imjangId: item.noteId, version: 0)
+                    return
+                }
+                
                 callVersionRequest(imjangId: item.noteId) { [weak self] version in
                 guard let self else { return }
                     if let version = version {
@@ -415,6 +467,12 @@ extension ImjangListViewController: UICollectionViewDataSource, UICollectionView
                 }
             case .list:
                 let imjangId = imjangList[indexPath.row].noteId
+                // 온보딩 분기처리
+                if UserDefaultManager.shared.isOnboarding {
+                    self.showImjangNoteVC(imjangId: imjangId, version: 0)
+                    return
+                }
+                
                 callVersionRequest(imjangId: imjangId) { [weak self] version in
                     guard let self else { return }
                     if let version = version {
@@ -429,7 +487,9 @@ extension ImjangListViewController: UICollectionViewDataSource, UICollectionView
         }
     }
     
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+    func collectionView(_ collectionView: UICollectionView,
+                        viewForSupplementaryElementOfKind kind: String,
+                        at indexPath: IndexPath) -> UICollectionReusableView {
         if let section = Section(rawValue: indexPath.section) {
             switch section {
             case .scrap:
@@ -450,6 +510,8 @@ extension ImjangListViewController: UICollectionViewDataSource, UICollectionView
                     header.shareButton.addTarget(self, action: #selector(showShareSelectVC), for: .touchUpInside)
                     header.filterActionRelay
                         .subscribe(with: self) { owner, action in
+                            // 온보딩 분기처리
+                            guard (!UserDefaultManager.shared.isOnboarding) else { return }
                             print(action)
                             switch action {
                             case .updated:
