+++ juinjang/Projects/App/Sources/UserInterface/RecordingRoom/ImjangImageList/ImjangImageListViewController.swift
@@ -43,7 +43,18 @@ final class ImjangImageListViewController: BaseViewController {
     private var isDeleteMode = false
     
     private lazy var panGesture = UIPanGestureRecognizer(target: self, action: #selector(handlePanGesture(_:)))
-
+    
+    init(images: [String] = []) {
+        self.imageList = images.enumerated().map { index, url in
+            ImageDto(imageId: index, imageUrl: url)
+        }
+        super.init()
+    }
+    
+    @MainActor required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         bindNavigationBar()
@@ -58,6 +69,7 @@ final class ImjangImageListViewController: BaseViewController {
     
     // 이미지 전체 조회 요청
     func callFetchImageRequest() {
+        guard (!UserDefaultManager.shared.isOnboarding) else { return }
         guard let imjangId = imjangId else { return }
         JuinjangAPIManager.shared.fetchData(type: BaseResponse<ImagesListDto>.self, api: .fetchImage(imjangId: imjangId)) { response, error in
             if let error = error {
@@ -174,8 +186,16 @@ final class ImjangImageListViewController: BaseViewController {
                 case .popButtonTap:
                     self.popView()
                 case .trashButtonTap:
+                    if UserDefaultManager.shared.isOnboarding {
+                        self.present(SignUpBottomSheetView(), animated: true)
+                        return
+                    }
                     self.deleteImageButtonTapped()
                 case .addButtonTap:
+                    if UserDefaultManager.shared.isOnboarding {
+                        self.present(SignUpBottomSheetView(), animated: true)
+                        return
+                    }
                     self.addImage()
                 default: break
                 }
@@ -328,7 +348,8 @@ extension ImjangImageListViewController: UICollectionViewDelegate, UICollectionV
 
 // MARK: ImagePicker Delegate
 extension ImjangImageListViewController: PHPickerViewControllerDelegate {
-    func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
+    func picker(_ picker: PHPickerViewController,
+                didFinishPicking results: [PHPickerResult]) {
         picker.dismiss(animated: true)
         let group = DispatchGroup()
